[SofaQtQuickGUI] WIP begin to add a new panel to for rendering the sofa camera.
The target of this branch is to separate in two panels the
3D view.

The first panel is the existing SofaViewer. It implement
the Editor's views in which we can manipulate the scene and visualize it
using the editor's camera.

The second panel is new RuntimeViewer. It will implement
the rendering of the in-scene cameras. These cameras
possibly be only controlled using in-scene interaction
components or python scripts but not with the editor's interactors.
All the key&mouse done in this panel are forwarded
as Sofa events.

TO BE CONTINUED