[runSofa2] Add a ViewSelector to replace the DynamicContent
DynamicContent scans a directory to define the content of the menu.
The menu entries are defined the file name storing the view.
The menu ordering is defined by the file ordering.

This is nice while developping new user interfaces but for the final version
it is important to be able to have:
  - application specific names like "3D View" (an not coder's name like SofaOpenGLViewer)
  - order of entries should stay constant and controlled to have the most important
    entries at the extremities.

This is what ViewSelector implements.

The component also expose the qml files in a directory but with
"_"+SofaViewer+"_" with the hope it will be clearer
this is only for developement version.