[SofaQtQuickGUI] Split SofaApplication in two.
SofaApplication was an "all-in-one" class which make extensibility not
very obivious and the interactions with the "OurInstance" looks unusual.

This commit is an attempt to clean that by splitting the class in two.

One class that I named MainApplication contains the old parts that are needed by SofaApplication.qml
This class have static function forwarding to the application specific
implementation.

The forwarding is done through a base class called SofaApplication.
The old implementation is now in DefaultApplication (that inherit from SofaApplication).

This is more a WIP to see if this helps in clarifying things.