RENAME ControlledSofaView to CameraView and make it display the BaseCamera instead of InteractiveCamera
Most of the time, interaction are "application" specific.
So we should implement the movement of the camera and the connection to
user input as external code (eg: PythonScriptController).

So it better the widget to display BaseCamera so that anyone can take profit of this
view to render the "in-scenes" cameras.