ADD CameraView and EditView
The two are different implementation of the 3D view.
CameraView is for in scene camera rendering, all the events should be passed
to the Sofa pipeline. We can only have one CameraView (otherwise it will be a
giant mess to forward the user inputs through the scene graph).

EditView is the 'editors' view where we can interact with editor's tools.