RENAME SofaWidget into SofaView
'Views' are integrating a lot of functionnality into a top level panel
that can be displayed as is to the user eg: Inspector, CameraView, EditView,
SceneGraphView. A view is connected to the workflow of user and how they set they
their workspace.

Widget is a more simple element, a single widget can be integrated into multiple views.
Eg of widget: Header, SearchBar,