ADD multilink autojump & signalling to SofaSceneGraphView
All the icons in the rows are now taking the same spacing to avoid having
zig-zagging hierarhcy.

The rendering is changed for multi-parent node to allow to differentiate them more easily.
Double clikcing on a multiparent node "auto jump" the selected object to the
implementation.

Implement a way to visually signal (by blinking) that the selection
is changing.




 the first instance is rendered differently, by double clikcing on
a multi-parent node it is now possible to "autojump" to its "details".