[SofaQtQuickGUI] ADD SofaViewListModel in SofaQtQuickGUI.cpp
To expose a c++ object into QML it is needed to register it into the
qml factory system. This is what is done in this commit.

As the SofaViewListModel is supposed to expose a single and consistent
view of the dedicated UI directory it is a singleton so each view is not
repeating always the same similar operations.