[SofaQtQuick] REFACTORING of how the dynamic update of SofaData's view is working
The old system was connecting the QmlDDGNode to the widget. So when
a datafield was changed this has the consequence to trigger the update
of some widget which can have as a consequence to pull other data and a
weird way.
To avoid this problem we are now using a timer to trigger the data widget
at regular interval.