[runSofa2] Fix compilation issue related to missing runSofa2.h
To improve containment of plugins and libraries it was decided to stop
including a default directory but package them in separate location for
each plugin. As Sofa is no setting the directory "include" by default we
need to implement the recommanded solution in the plugin...this is what this
PR does.